Fix Android parameters ordering
The current implementation raises "cannot have non-function argument after a function arg"
as of RN v0.48+.

This patch fixes this issue inverting the order of the arguments in the native implementation.
Proper checks are in place to make this change backward compatible and detect the previous API usage,
switching the parameters as needed.